+++ lib/classes/humanClass.py
@@ -1,11 +1,10 @@
-import random
-import json
+import random, json, sqlite3
 from lib.variables.main import foods, foods_energy, shop_items
 
 
 class Human:
     def __init__(
-        self, name, coins=10, health=100, hunger=0, thirst=0, energy=100, inventory=None
+        self, name, coins=10, health=100, hunger=0, thirst=0, energy=100, foods_inventory=None, items_inventory=None
     ):
         self.name = name
         self.coins = coins
@@ -13,7 +12,8 @@ class Human:
         self.hunger = hunger
         self.thirst = thirst
         self.energy = energy
-        self.inventory = inventory if inventory is not None else []
+        self.foods_inventory = foods_inventory if foods_inventory is not None else []
+        self.items_inventory = items_inventory if items_inventory is not None else []
         self.xp = 0
         self.level = 1
 
@@ -22,88 +22,92 @@ class Human:
         if food_found:
             food = random.choice(foods)
             print(f"{self.name} found some {food}.")
-            self.energy -= random.randint(5, 15)
-            self.hunger += random.randint(5, 15)
-            self.thirst += random.randint(5, 15)
-            self.health -= random.randint(5, 10)
-            self.inventory.append(food)
+            self.foods_inventory.append(food)
             self.gain_xp(7)
         else:
             print(f"{self.name} found no food.")
-            self.energy -= 10
-            self.hunger += 10
-            self.thirst += 10
-            self.health -= random.randint(5, 10)
+            self.update_stats('bad', random.randint(1, 10), random.randint(1, 10), random.randint(1, 10), random.randint(1, 10))
 
     def eat(self, food):
         food_energy = foods_energy
-        if food in self.inventory:
-            self.health = min(self.health + 5, 100)
-            self.hunger = max(self.hunger - food_energy[food], 0)
-            self.thirst = max(self.thirst - 3, 0)
-            self.energy = min(self.energy + food_energy[food], 100)
-            self.inventory.remove(food)
-            print(f"{self.name} eats {food} and gains {food_energy[food]} energy.")
+        if food in self.foods_inventory:
+            self.foods_inventory.remove(food)
+            self.update_stats('good', min(self.health + 5, 100), max(self.hunger - food_energy[food], 0), max(self.thirst - 3, 0), min(self.energy + food_energy[food], 100))
             self.gain_xp(5)
+            print(f"{self.name} eats {food} and gains {food_energy[food]} energy.")
         else:
             print(f"{self.name} doesn't have any {food} to eat.")
 
     def drink(self):
         print(f"{self.name} drinks water.")
-        self.health = min(self.health + 10, 100)
-        self.thirst = max(self.thirst - 12, 0)
-        self.energy = min(self.energy + 5, 100)
-        self.gain_xp(3)
+        self.update_stats('good', 5, 5, 15, 10)
+        self.gain_xp(2)
 
     def sleep(self, hours):
         if hours > 8:
             print(f"Caution! {self.name} sleep for >8 hours!")
-            self.energy = min(self.energy + 10 * 3, 100)
-            self.hunger += 15
-            self.thirst += 21
+            self.update_stats('neutral', 10, random.randint(5, 15), random.randint(5, 20), 25 )
+            self.gain_xp(3 * 10)
         else:
           print(f"{self.name} sleep for {hours} hours.")
-          self.energy = min(self.energy + hours * 3, 100)
-          self.hunger += 5
-          self.thirst += 7
+          self.update_stats('neutral', 5, random.randint(1, 10), random.randint(5, 10), hours * 3)
           self.gain_xp(3 * hours)
 
     def shop(self, item):
-        inventory = shop_items
-        if item in inventory and self.coins >= inventory[item]:
-            self.coins -= inventory[item]
-            self.inventory.append(item)
-            print(f"{self.name} bought {item} for {inventory[item]} coins.")
+        items = shop_items
+        if item in items and self.coins >= items[item]:
+            self.items_inventory.append(item)
+            self.coins -= items[item]
+            print(f"{self.name} bought {item} for {items[item]} coins.")
         else:
             print(f"{self.name} can't afford {item}.")
 
     def chop_tree(self):
-        if "Axe" in self.inventory:
+        if "Axe" in self.items_inventory:
+            found_gapple = random.choice([True, False])
             print(f"{self.name} is chopping a tree...")
-            self.energy -= 40
-            self.hunger += 30
-            self.thirst += 25
+            if found_gapple:
+                print(f"Lucky! {self.name} found 1 Golden Apple when chopping a tree!")
+                self.foods_inventory.append('Golden Apple')
+            else:
+                print(f"{self.name} chopped a tree and earned some coins.")
+            self.update_stats('bad', 5, random.randint(5, 30), random.randint(5, 25), 40)
             self.coins += random.randint(1, 75)
-            print(f"{self.name} chopped a tree and earned 1 - 75 coins.")
             self.gain_xp(10)
         else:
             print(f"{self.name} needs an Axe to chop a tree.")
-
-    """
-    TODO: improve this
-    
-    def lucky_box(self):
-        print("Opening lucky box with 10 coins...")
-        self.coins -= 10
-        win = random.choice([True, False])
-        if win:
-            print("Nice, You win!")
-            self.health = min(self.health + 25, 100)
-            self.hunger = 0
+            
+    def mining(self):
+        if "Pickaxe" in self.items_inventory:
+            found_diamond = random.choice([True, False])
+            print(f"{self.name} is mining...")
+            if found_diamond:
+                self.update_stats('bad', random.randint(5, 50), random.randint(5, 25), random.randint(5, 25), random.randint(5, 50))
+                self.coins += 150
+                print(f"Lucky! {self.name} found diamond and earned 150 coin!")
+            else:
+                self.update_stats('bad', random.randint(1, 25), random.randint(1, 15), random.randint(1, 15), random.randint(1, 25))
+                self.coins += random.randint(10, 100)
+                print(f"{self.name} mined and earned some coins.")
         else:
-            print("Loser. You lose everything, including your life :)")
-            self.health = 0
-    """
+            print(f"{self.name} needs a Pickaxe to mine.")
+
+    def update_stats(self, mode, health, hunger, thirst, energy):
+        if mode == 'good':
+            self.health += health
+            self.hunger -= hunger
+            self.thirst -= thirst
+            self.energy += energy
+        elif mode == 'bad':
+            self.health -= health
+            self.hunger += hunger
+            self.thirst += thirst
+            self.energy -= energy
+        elif mode == 'neutral':
+            self.health += health
+            self.hunger += hunger
+            self.thirst += thirst
+            self.energy += energy
 
     def gain_xp(self, amount):
         self.xp += amount
@@ -126,11 +130,12 @@ class Human:
             print(f"{self.name} has died")
             exit()
         else:
+            if self.hunger < 1:
+                self.hunger = 25
+            elif self.thirst < 1:
+                self.thirst = 30
             print(f"{self.name} is surviving.")
-
-    def show_inventory(self):
-        print(f"{self.name}'s inventory:\n {self.inventory}")
-
+            
     def save_game(self, filename="savegame.json"):
         state = {
             "name": self.name,
@@ -139,7 +144,8 @@ class Human:
             "hunger": self.hunger,
             "thirst": self.thirst,
             "energy": self.energy,
-            "inventory": self.inventory,
+            "foods_inventory": self.foods_inventory,
+            "items_inventory": self.items_inventory,
             "xp": self.xp,
             "level": self.level,
         }
@@ -156,7 +162,25 @@ class Human:
         self.hunger = state["hunger"]
         self.thirst = state["thirst"]
         self.energy = state["energy"]
-        self.inventory = state["inventory"]
+        self.foods_inventory = state["foods_inventory"]
+        self.items_inventory = state["items_inventory"]
         self.xp = state["xp"]
         self.level = state["level"]
         print(f"Game loaded from {filename}")
+
+
+    """
+    TODO: improve this
+    
+    def lucky_box(self):
+        print("Opening lucky box with 10 coins...")
+        self.coins -= 10
+        win = random.choice([True, False])
+        if win:
+            print("Nice, You win!")
+            self.health = min(self.health + 25, 100)
+            self.hunger = 0
+        else:
+            print("Loser. You lose everything, including your life :)")
+            self.health = 0
+    """