+++ docs/changelog.md
@@ -1,15 +1,22 @@
 # Changelog
 
-## v0.1.2 -
+## v0.2.0 -  (Major, Latest)
 
-* Updated: When player found foods, some stats will decreased
+* Added: Documentation (readme, changelog, contribute)
+* Added: mining (Pickaxe required)
+* Added: ```update_stats()``` method to make code readable
+* Added: random event when chopping tree and mining (found golden apple, found diamond)
+* Updated: Using ```default_name``` when player not submitting a name
+* Updated: Using ```random.randint``` for increasing/decreasing some player stats
+* Fixed: Player stats not changed when found foods
+* Fixed: Hunger and Thirst is negative number
 * Fixed: Player not die when thirst or hunger is >100
 
 ## v0.1.1 - Oct, 7 2024
 
 * Added: XP & Leveling System
 * Fixed: User interface
-* Updated: Sleeping for >8 hours now increasing 30 of energy, but increasing 2x of thirst and hunger than <8 hours 
+* Updated: Warning when player trying to sleep for >8 hours
 
 ## v0.1.0